rename (post) : 북마크 관련 파일 post 도메인 내로 이동
- 사실상 bookmark 는 게시글 / 구인구직 글을 그대로 반환하므로 post에 조금더 가깝다고 판단하여 이동
