+++ src/main/java/com/backend/immilog/post/application/command/PostUploadCommand.java
@@ -1,6 +1,6 @@
 package com.backend.immilog.post.application.command;
 
-import com.backend.immilog.post.domain.enums.Categories;
+import com.backend.immilog.post.domain.model.enums.Categories;
 import io.swagger.annotations.ApiModel;
 import lombok.Builder;
