+++ src/main/java/com/backend/immilog/post/application/result/CommentResult.java
@@ -1,7 +1,7 @@
 package com.backend.immilog.post.application.result;
 
 import com.backend.immilog.post.domain.model.Comment;
-import com.backend.immilog.post.domain.enums.PostStatus;
+import com.backend.immilog.post.domain.model.enums.PostStatus;
 import com.backend.immilog.user.application.result.UserInfoResult;
 import com.backend.immilog.user.domain.model.User;
 import lombok.AllArgsConstructor;
