+++ src/main/java/com/backend/immilog/post/application/services/PostInquiryService.java
@@ -3,9 +3,9 @@ package com.backend.immilog.post.application.services;
 import com.backend.immilog.post.exception.PostException;
 import com.backend.immilog.post.application.result.CommentResult;
 import com.backend.immilog.post.application.result.PostResult;
-import com.backend.immilog.post.domain.enums.Categories;
-import com.backend.immilog.post.domain.enums.Countries;
-import com.backend.immilog.post.domain.enums.SortingMethods;
+import com.backend.immilog.post.domain.model.enums.Categories;
+import com.backend.immilog.post.domain.model.enums.Countries;
+import com.backend.immilog.post.domain.model.enums.SortingMethods;
 import com.backend.immilog.post.domain.repositories.CommentRepository;
 import com.backend.immilog.post.domain.repositories.PostRepository;
 import lombok.RequiredArgsConstructor;
