+++ src/main/java/com/backend/immilog/post/application/services/PostUpdateService.java
@@ -5,7 +5,7 @@ import com.backend.immilog.post.application.command.PostUpdateCommand;
 import com.backend.immilog.post.exception.PostException;
 import com.backend.immilog.post.domain.model.InteractionUser;
 import com.backend.immilog.post.domain.model.Post;
-import com.backend.immilog.post.domain.enums.ResourceType;
+import com.backend.immilog.post.domain.model.enums.ResourceType;
 import com.backend.immilog.post.domain.repositories.BulkInsertRepository;
 import com.backend.immilog.post.domain.repositories.InteractionUserRepository;
 import com.backend.immilog.post.domain.repositories.PostRepository;
@@ -20,10 +20,10 @@ import java.util.List;
 import java.util.Objects;
 
 import static com.backend.immilog.post.exception.PostErrorCode.*;
-import static com.backend.immilog.post.domain.enums.InteractionType.LIKE;
-import static com.backend.immilog.post.domain.enums.PostType.POST;
-import static com.backend.immilog.post.domain.enums.ResourceType.ATTACHMENT;
-import static com.backend.immilog.post.domain.enums.ResourceType.TAG;
+import static com.backend.immilog.post.domain.model.enums.InteractionType.LIKE;
+import static com.backend.immilog.post.domain.model.enums.PostType.POST;
+import static com.backend.immilog.post.domain.model.enums.ResourceType.ATTACHMENT;
+import static com.backend.immilog.post.domain.model.enums.ResourceType.TAG;
 
 @Slf4j
 @Service
