+++ src/main/java/com/backend/immilog/post/domain/model/Post.java
@@ -1,8 +1,8 @@
 package com.backend.immilog.post.domain.model;
 
 import com.backend.immilog.post.application.command.PostUploadCommand;
-import com.backend.immilog.post.domain.enums.Categories;
-import com.backend.immilog.post.domain.enums.Countries;
+import com.backend.immilog.post.domain.model.enums.Categories;
+import com.backend.immilog.post.domain.model.enums.Countries;
 import com.backend.immilog.post.domain.vo.PostMetaData;
 import com.backend.immilog.post.domain.vo.PostUserData;
 import com.backend.immilog.user.domain.model.User;
