+++ src/main/java/com/backend/immilog/post/domain/model/PostResource.java
@@ -1,7 +1,7 @@
 package com.backend.immilog.post.domain.model;
 
-import com.backend.immilog.post.domain.enums.PostType;
-import com.backend.immilog.post.domain.enums.ResourceType;
+import com.backend.immilog.post.domain.model.enums.PostType;
+import com.backend.immilog.post.domain.model.enums.ResourceType;
 import lombok.Builder;
 
 @Builder
