+++ src/main/java/com/backend/immilog/post/domain/model/enums/Categories.java
@@ -1,4 +1,4 @@
-package com.backend.immilog.post.domain.enums;
+package com.backend.immilog.post.domain.model.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
