+++ src/main/java/com/backend/immilog/post/domain/model/enums/SortingMethods.java
@@ -1,4 +1,4 @@
-package com.backend.immilog.post.domain.enums;
+package com.backend.immilog.post.domain.model.enums;
 
 public enum SortingMethods {
     CREATED_DATE,
