+++ src/main/java/com/backend/immilog/post/domain/repositories/BookmarkRepository.java
@@ -1,0 +1,9 @@
+package com.backend.immilog.post.domain.repositories;
+
+import com.backend.immilog.post.domain.model.Bookmark;
+
+public interface BookmarkRepository {
+    void saveEntity(
+            Bookmark bookmark
+    );
+}
