+++ src/main/java/com/backend/immilog/post/domain/repositories/PostRepository.java
@@ -2,9 +2,9 @@ package com.backend.immilog.post.domain.repositories;
 
 import com.backend.immilog.post.application.result.PostResult;
 import com.backend.immilog.post.domain.model.Post;
-import com.backend.immilog.post.domain.enums.Categories;
-import com.backend.immilog.post.domain.enums.Countries;
-import com.backend.immilog.post.domain.enums.SortingMethods;
+import com.backend.immilog.post.domain.model.enums.Categories;
+import com.backend.immilog.post.domain.model.enums.Countries;
+import com.backend.immilog.post.domain.model.enums.SortingMethods;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
