+++ src/main/java/com/backend/immilog/post/infrastructure/repositories/PostResourceRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.backend.immilog.post.infrastructure.repositories;
 
-import com.backend.immilog.post.domain.enums.ResourceType;
+import com.backend.immilog.post.domain.model.enums.ResourceType;
 import com.backend.immilog.post.domain.repositories.PostResourceRepository;
 import com.backend.immilog.post.infrastructure.jpa.entities.QPostResourceEntity;
 import com.querydsl.core.types.dsl.BooleanExpression;
