+++ src/main/java/com/backend/immilog/post/presentation/request/PostUploadRequest.java
@@ -1,7 +1,7 @@
 package com.backend.immilog.post.presentation.request;
 
 import com.backend.immilog.post.application.command.PostUploadCommand;
-import com.backend.immilog.post.domain.enums.Categories;
+import com.backend.immilog.post.domain.model.enums.Categories;
 import io.swagger.annotations.ApiModel;
 import lombok.Builder;
