--- src/main/java/com/backend/immilog/user/domain/repositories/BookmarkRepository.java
@@ -1,9 +1,0 @@
-package com.backend.immilog.user.domain.repositories;
-
-import com.backend.immilog.user.domain.model.Bookmark;
-
-public interface BookmarkRepository {
-    void saveEntity(
-            Bookmark bookmark
-    );
-}
